pyproject.toml
[tool.poetry]
name = "prvaite"
version = "0.1.0"
description = ""
authors = ["artemi <artemi13221@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.1.3"
python-dotenv = "^0.19.2"
webdriver-manager = "^3.5.4"
schedule = "^1.1.0"
DateTime = "^4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
